Update audit.md
Remove/change audit question "Try passing as arguments `--color=orange "HeY GuYs"`, in order to color `GuYs`." since task description requires the program to be able to alter the color of specified letters, not words. Currently, the program colors "Y" in "HeY" in addition to "GuYs".
Implementing the ability to specify words to be colored should be a bonus task as is multiple colors per string and different color flag notations.